Added chat interface and schema. Also added a route to post new chats.
Chats hold the user ids of all users in them, a boolean for if they are a groupchat, an array of strings for messages and a chatname. To post a chat use http://localhost:8000/chat/create. Format the body as follows:

{
   "chatName": "JsonTestChat",
   "isGroupChat": true,
   "users": ["65515fda44de0514c4f8bc80", "65514aeda5f5a90678cb5952"],
   "messages": ["First message", "Second message"]
}

This post body is in Json format. Must have bearer token to post chats.